Excract 409 saving error (conflict) to a separate exception. Sometimes it's necessary to handle them in a special way.
Sometime users perform double click (and you can not change it)
and you get a double response. In this situation you can have 409
error in one of the requests.

It's quite safe to ignore it, but it will float your logs and
admin emails. So probably you will want to catch it and ignore.

```
try:
    o.save()
except SyncGatewayConflict:
    pass
```